Search for a task and add an optional due date for a new created task

+++ src/main/webapp/app/tasks/task-list/task-list.component.ts
@@ -2,6 +2,7 @@ import { ChangeDetectionStrategy, Component, EventEmitter, Inject, Input, Output
 
 import { Task } from '../task';
 import { TaskService } from '../task.service';
+import Swal from 'sweetalert2'
 
 /**
  * A list of tiny tasks.
@@ -14,15 +15,38 @@ import { TaskService } from '../task.service';
 })
 export class TaskListComponent {
 
-  @Input() tasks: Task[] | null = null;
+  @Input() tasks: Task[] | null = [];
 
   @Output() deleted: EventEmitter<Task> = new EventEmitter();
 
+  searchValue: any = '';
+
   constructor(@Inject('TaskService') private taskService: TaskService) {}
 
+  searchTask(event: string) {
+    // @ts-ignore
+    this.searchValue = event;
+  }
+
   delete(task: Task): void {
-    this.taskService.delete(task.id).subscribe(() => {
-      this.deleted.emit(task);
-    });
+    Swal.fire({
+      title: 'Do you want to delete this task?',
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      denyButtonText: `No`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        this.taskService.delete(task.id).subscribe(() => {
+          this.deleted.emit(task);
+          Swal.fire('Task has been deleted!', '', 'success');
+
+        });
+
+      } else if (result.isDenied) {
+        Swal.fire('The task was not deleted', '', 'info')
+      }
+    })
   }
 }
